+++ packages/json_schema_builder/lib/src/schema_cache_file.dart
@@ -4,13 +4,8 @@
 
 import 'dart:io';
 
-import 'schema_cache_base.dart';
-
-class SchemaCache extends SchemaCacheBase {
-  SchemaCache({super.httpClient, super.loggingContext});
-
-  @override
-  Future<String> getCachedFile(Uri uri) async {
+class SchemaCacheFileLoader {
+  Future<String> getFile(Uri uri) async {
     assert(uri.scheme == 'file');
     final file = File.fromUri(uri);
     return file.readAsString();
