+++ packages/json_schema_builder/lib/src/schema_cache_web.dart
@@ -2,13 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'schema_cache_base.dart';
-
-class SchemaCache extends SchemaCacheBase {
-  SchemaCache({super.httpClient, super.loggingContext});
-
-  @override
-  Future<String> getCachedFile(Uri uri) async {
+class SchemaCacheFileLoader {
+  Future<String> getFile(Uri uri) async {
     throw UnimplementedError(
       'file:// schemes not supported for schema cache on web.',
     );
